requirements.txt
Django>=4.2.0
pyzk>=0.9
python-dotenv>=1.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-q>=1.3.9
openpyxl>=3.1.0
